Updated deploy.sh

+++ scripts/deploy.sh
@@ -4,8 +4,10 @@ cd ~/amazon-redshift-infrastructure-automation
 python3 -m venv .env
 source .env/bin/activate
 pip install -r requirements.txt
-echo '~/amazon-redshift-infrastructure-automation/scripts/restart_session.sh'  >> ~/.bashrc
-read -n 1 -r -s -p $"Upload user-config.json and press enter to continue...\n"
+LINE='~/amazon-redshift-infrastructure-automation/scripts/restart_session.sh'
+FILE='~/.bashrc'
+grep -qF -- "$LINE" "$FILE" || echo "$LINE" >> "$FILE"
+read -n 1 -r -s -p $'Upload user-config.json and press enter to continue...\n'
 #Need more elegant solution for handling exception here:
-mv ~/user-config.json ~/amazon-redshift-infrastructure-automation/
+[ -f ~/user-config.json ] && mv ~/user-config.json ~/amazon-redshift-infrastructure-automation/user-config.json
 cdk deploy --all --require-approval never
